use deployment status test suite
