+++ app-deployment-manager/test/deployment/methods_test.go
@@ -82,7 +82,7 @@ func (s *TestSuite) TestAPIMethods() {
 			methodMap:   methodResponses["deploymentById"],
 			description: "get and delete deployment",
 			setup: func() string {
-				deployID, retCode, err := utils.StartDeployment(admclient, dpConfigName, "targeted", 10)
+				deployID, retCode, err := utils.StartDeployment(admclient, utils.AppNginx, "targeted", utils.DeploymentTimeout, "GetAndDeleteDeployment")
 				s.Equal(retCode, http.StatusOK)
 				s.NoError(err)
 				return deployID
@@ -100,7 +100,7 @@ func (s *TestSuite) TestAPIMethods() {
 			methodMap:   methodResponses["deploymentClusters"],
 			description: "list deployment clusters",
 			setup: func() string {
-				deployID, retCode, err := utils.StartDeployment(admclient, dpConfigName, "targeted", 10)
+				deployID, retCode, err := utils.StartDeployment(admclient, utils.AppNginx, "targeted", utils.DeploymentTimeout, "ListDeploymentClusters")
 				s.Equal(retCode, http.StatusOK)
 				s.NoError(err)
 				return deployID
