+++ app-deployment-manager/test/deployment/negative_test.go
@@ -29,7 +29,7 @@ func (s *TestSuite) TestNegativeCreateDeployment() {
 		err := ResetThenChangeDpConfig("nginx", test.configKey, test.configValue, originalDpConfigs)
 		s.NoError(err, "failed to reset "+test.configKey+" in deployment config")
 
-		deployID, retCode, err := utils.StartDeployment(s.AdmClient, "nginx", test.deployment, 10)
+		deployID, retCode, err := utils.StartDeployment(s.AdmClient, "nginx", test.deployment, utils.DeploymentTimeout, "TestNegativeCreateDeployment")
 		s.Equal(retCode, http.StatusBadRequest)
 		s.Error(err)
 		s.Contains(err.Error(), test.expectedErr)
